requirements.txt
gitpython==3.1.42
numpy>=1.23.0
scikit-learn>=1.2.0
nltk>=3.8.0
requests>=2.28.0
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
# Optional dependencies
pdfkit>=1.0.0
markdown>=3.3.0
wkhtmltopdf-pack==0.12.3.0.post1
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 